moves wordclouds in own directory, also wc code in its own module
