--- .env-example.txt
@@ -1,10 +1,0 @@
-
-# google cloud
-GCP_PROJECT=
-GCP_REGION=
-DOCKER_REPO_NAME=
-GAR_MEMORY=
-
-# Docker
-DOCKER_IMAGE_NAME=cyb
-DOCKER_LOCAL_PORT=8080
