+++ .env_docker
@@ -9,5 +9,5 @@ DOCKER_IMAGE_PATH=gcr.io/$GCP_PROJECT/$APP_NAME
 
 MODEL_PATH=/models/model.pkl
 
-WORDCLOUD_REAL_PATH=/models/words_real.json
-WORDCLOUD_FAKE_PATH=/models/words_fake.json
+WORDCLOUD_REAL_PATH=/wordcloud/words_real.json
+WORDCLOUD_FAKE_PATH=/wordcloud/words_fake.json
