+++ .version
@@ -1,1 +1,1 @@
-0.0.42
+0.0.44
