+++ wordcloud/words_fake.json
@@ -1,0 +1,1 @@
+{"hillary": 0.0024683022731492804, "clinton": 0.001857431339378012, "video": 0.0015389522876528452, "image": 0.0012798818037452137, "get": 0.0011969692589686907, "\u2013": 0.0010906173332594963, "go": 0.001090390650962301, "via": 0.001085710573969532, "know": 0.0010483377377802877, "obama": 0.0010020251889717655, "america": 0.0009913653106418535, "people": 0.0009589377396697365, "fbi": 0.000940617450721505, "medium": 0.0008896322381178614, "no": 0.0008784248574003517, "email": 0.0008419392504387366, "like": 0.0008335864175462002, "american": 0.0008086429637578639, "donald": 0.0007794016101244164, "fact": 0.0007719253399737695, "even": 0.0007578020870186886, "only": 0.0007474791905017703, "black": 0.0006998403426564233, "world": 0.0006647447775122405, "war": 0.0006554565163848368, "thing": 0.0006363429936162415, "lie": 0.0005574069890242516, "one": 0.0005567834232909404, "trump": 0.0005409112727751304, "claim": 0.0005281943724236747, "live": 0.0005219369904894495, "watch": 0.0005217785676743937, "think": 0.0005108748236198153, "october": 0.0005108482018574428, "actually": 0.0005078044167945934, "de": 0.0005045021577588217, "news": 0.000497130093492823, "see": 0.00047782385943327373, "use": 0.00047730888041850705, "story": 0.0004726431464570322, "should": 0.0004693737230278556, "post": 0.00046668354402725376, "right": 0.0004586751009625831, "woman": 0.0004488594841022494, "really": 0.000444885881018564, "\u2018": 0.00043785240952848336, "happen": 0.00043153204520193084, "make": 0.000429271057374947, "every": 0.0004204268191585858, "man": 0.00040466635969204343, "la": 0.00039142550702929576, "today": 0.0003905649923683493, "article": 0.00038824967006894924, "racist": 0.00038151121478376875, "never": 0.0003731652490003963, "not": 0.0003729713264955803, "let": 0.00037078237668997343, "november": 0.0003597299552062291, "nothing": 0.0003581629300380639, "truth": 0.0003514279631781084, "realdonaldtrump": 0.00034960342825331213, "believe": 0.0003492549542238702, "read": 0.0003459191632001198, "wire": 0.0003423294344599012, "way": 0.00033794762818941987, "tweet": 0.0003318283303948911, "anyone": 0.00033122549795634466, "course": 0.0003297731687924362, "ever": 0.0003289135203606955, "mean": 0.0003284612415312011, "come": 0.00032783059705999333, "reason": 0.00032374183416297334, "look": 0.0003228167026383731, "\u0432": 0.00032180017963001845, "wikileaks": 0.0003150191146080581, "word": 0.0003059831854700912, "reveal": 0.0003056346735601001, "cnn": 0.0002994193199584505, "matter": 0.00029484290794462055, "entire": 0.0002899415942567581, "explain": 0.00028961626403625003, "well": 0.0002881192098650629, "something": 0.00028724532814331395, "however": 0.0002855008636691777, "gun": 0.00028254932980098117, "muslim": 0.00027490762338669417, "anything": 0.00027211380681591577, "fox": 0.00026692978439852596, "destroy": 0.00026606328846566203, "give": 0.0002654903532504486, "foundation": 0.0002642054031549103, "real": 0.0002631622452944501, "fake": 0.00026277180565691523, "money": 0.00025989264026325036, "mainstream": 0.0002552943674619673, "great": 0.0002525403489760866, "gop": 0.00025221280811594117, "break": 0.0002518062158920698, "child": 0.00025166195450628513, "god": 0.00025058868384028084, "source": 0.0002498899968046774, "terrorist": 0.00024702763726229033, "\u0438": 0.0002465813880570488, "comey": 0.00024467936176116934, "stand": 0.00024301078293108476, "hate": 0.00024244504538442572, "liberal": 0.00024178861401781006, "information": 0.00024001111010906828, "century": 0.00023409158145293522, "want": 0.0002319754129075386, "everyone": 0.00023097965060271151, "isi": 0.0002304836305902703, "true": 0.0002273372727785673, "sure": 0.00022728123587444858, "person": 0.0002247386603454356, "cia": 0.00022409375907581793, "daily": 0.0002233458996273131, "election": 0.00022223387999005442, "love": 0.00022209683064497177, "russia": 0.00022185959293885603, "point": 0.00022113476232210583, "history": 0.00022077442635386088, "someone": 0.00021995526063420061, "cop": 0.00021967740434837026, "supporter": 0.00021960773906684887, "good": 0.00021909948082182544, "expose": 0.00021836174957129732, "st": 0.00021644168799088792, "evidence": 0.00021473523689054133, "prove": 0.0002144500353143214, "please": 0.00021382989009897685, "water": 0.00021355126557102592, "photo": 0.00021262606742158187, "free": 0.00021171145866124636, "secret": 0.00021085057817830696, "shoot": 0.00021001409706390196, "guy": 0.00020849321317418384, "seem": 0.00020672087004254254, "bad": 0.00020621783268115574, "attempt": 0.0002060376274576914, "russian": 0.00020556989242324387, "podesta": 0.00020166083434247514, "en": 0.00019974580680778603, "life": 0.00019884594349887268, "continue": 0.00019775424950035827, "que": 0.0001965490275418888, "admit": 0.00019645735775144016, "th": 0.00019605906066210208, "simply": 0.00019188948258070288, "system": 0.00019181404637358206, "mind": 0.00019116170182648952, "thank": 0.0001897613828537128, "apparently": 0.00018959240853009738, "much": 0.00018928523770755283, "dnc": 0.00018864385622540146, "report": 0.0001884282232260321, "soros": 0.0001879651642691517, "investigation": 0.000187785537831945, "reality": 0.00018626162018982594, "propaganda": 0.00018600123215646878, "clearly": 0.00018409059977470615, "kid": 0.00018356607920206, "gold": 0.00018308815196273838, "everything": 0.00018125295317014992, "show": 0.00018124477953110362, "completely": 0.0001809800590693028, "need": 0.00018060214638971283, "cause": 0.0001769945404585859, "share": 0.0001719970468501692, "stop": 0.00017167690615105626, "dc": 0.0001709556645645629, "crime": 0.00017080926178193143, "facebook": 0.0001700610913820767, "pretty": 0.0001665873913023454, "el": 0.00016646383313149876, "hand": 0.00016567970190387016, "exactly": 0.00016519053201049887, "press": 0.00016467067442568685, "flag": 0.00016315249896487244, "understand": 0.00015742176462799285, "refuse": 0.00015697908746291412, "absolutely": 0.00015634721617640088, "old": 0.00015618931465901287, "become": 0.0001550616454605771, "massive": 0.00015479154863946485, "agent": 0.00015343449005046262, "conspiracy": 0.00015234065727304794, "john": 0.0001519557154903116, "regard": 0.0001513591373944835, "society": 0.00015075300075068193, "earth": 0.00014789656760956925, "elite": 0.000147475782594185, "christian": 0.00014705651445500126, "dollar": 0.0001466087744999074, "remember": 0.00014649967957342785, "wrong": 0.00014604058198710497, "keep": 0.00014578783320539322, "learn": 0.00014527640272238784, "instead": 0.0001439715098001924, "try": 0.00014393505118315232}
