+++ wordcloud/words_real.json
@@ -1,0 +1,1 @@
+{"say": 0.011411451940708212, "mr": 0.005258252036211692, "state": 0.0017594935827812729, "new": 0.0014899962433634097, "republican": 0.0013426943227991392, "would": 0.0012938822276086052, "official": 0.0011713728831825368, "year": 0.0011499786284055358, "house": 0.0011013952588688407, "but": 0.0010758206889994962, "last": 0.0009855766554202475, "percent": 0.0009842943052947412, "unite": 0.0009821537308625028, "party": 0.0009499808332043185, "york": 0.0009471533788046365, "leader": 0.0008349376516744923, "week": 0.0008274572084714381, "minister": 0.0008218575735699652, "president": 0.0008092290531227335, "court": 0.0008024089111880747, "senate": 0.0007957012650898217, "tuesday": 0.0007862871750453979, "two": 0.0007641884191907518, "north": 0.0007491053425326301, "include": 0.0007460899770329946, "government": 0.0007440298414913555, "thursday": 0.0007310913803506068, "china": 0.000730468158266419, "tax": 0.0007206228061767349, "wednesday": 0.0007073027728009346, "meet": 0.000702230798065434, "washington": 0.0006864035685841904, "company": 0.0006741491235063067, "friday": 0.0006724740081264488, "month": 0.000670814181721668, "monday": 0.0006614128512074945, "plan": 0.0006523744174686344, "korea": 0.0006467019711740064, "could": 0.0006464077023207717, "rule": 0.0006317772559419902, "former": 0.0006233825606643392, "security": 0.0006112072649372668, "statement": 0.000562711215772483, "administration": 0.0005601331223343194, "group": 0.0005562212408474678, "add": 0.0005446107184507634, "trade": 0.0005392038954452767, "deal": 0.0005385705503990053, "committee": 0.0005096729137204409, "senator": 0.0005063161777731932, "u": 0.0004947923146827356, "also": 0.0004673969229791689, "city": 0.000462643843801849, "spokesman": 0.00045042945288956994, "south": 0.0004494149393107888, "chief": 0.0004376774160524014, "lawmaker": 0.0004239741589801289, "executive": 0.00042269247910965794, "islamic": 0.00041929432497833354, "representative": 0.0004184585258731645, "iran": 0.000409981561259417, "sunday": 0.000408682056450734, "whether": 0.00040759461911956513, "democratic": 0.0004072008896489225, "agency": 0.0004061881176648072, "among": 0.00040333605093232613, "issue": 0.00040139002999073966, "seek": 0.0003930858220870885, "foreign": 0.000384493225997088, "european": 0.0003794472654934075, "congress": 0.00037731946496356655, "three": 0.0003684482443456154, "eu": 0.00036835913926259626, "country": 0.0003662638307433447, "move": 0.0003601206156618684, "prime": 0.0003568042879480174, "expect": 0.00035500541842314236, "sanction": 0.00035467597525676276, "budget": 0.00034693597888731597, "decision": 0.0003449858449998582, "senior": 0.00034446469787525764, "britain": 0.00034040476738024733, "lead": 0.0003394312630471077, "union": 0.0003377352458601947, "tell": 0.00033628478373991764, "national": 0.00032009511215618457, "campaign": 0.00031230956114838837, "policy": 0.00031039478445851255, "measure": 0.00030813567804730793, "hold": 0.00030409333339989685, "program": 0.00030148950270107505, "breitbart": 0.00030100258394583665, "legislation": 0.0003004957445899283, "remain": 0.00029960695932313553, "authority": 0.00029757619970716153, "governor": 0.0002967298229670849, "adviser": 0.0002953837391972247, "democrat": 0.000293495202793167, "office": 0.0002916003088234125, "reform": 0.0002887530318101457, "billion": 0.0002885244341701013, "opposition": 0.00028841311161389034, "member": 0.00028818586362017065, "saturday": 0.00028320620408622, "since": 0.00028310893845388775, "region": 0.00028008044888442306, "cut": 0.00027956809468362217, "agreement": 0.0002775818157484207, "presidential": 0.0002749431086511888, "business": 0.000273836781560485, "reporter": 0.0002720029010519798, "parliament": 0.00027155351487135877, "myanmar": 0.0002678991692674865, "next": 0.0002663329698300316, "thats": 0.00026427385075738027, "decline": 0.00026408071353789677, "military": 0.00026245311782710555, "capital": 0.0002598466191978758, "later": 0.00025863118542649154, "militant": 0.0002578670897767124, "visit": 0.00025559526956964624, "dont": 0.0002555284303121544, "ally": 0.0002495589296387579, "coalition": 0.0002492774699505232, "nomination": 0.0002471945877841848, "set": 0.0002468780936039185, "challenge": 0.0002466002347001415, "aid": 0.000245917104474523, "brief": 0.00024364960755964634, "turkey": 0.00024334971540264735, "mexico": 0.00024312182589818722, "chinese": 0.0002427997855107869, "ban": 0.00024126767072121992, "immigration": 0.00023924756060089074, "im": 0.00023642572426907973, "secretary": 0.00023637653748403086, "proposal": 0.00023610880154233756, "ministry": 0.0002353038286141901, "urge": 0.00023446903684175448, "lawyer": 0.000231582189080002, "effort": 0.00023136344656713348, "border": 0.0002311145285242382, "department": 0.00023089235507473438, "early": 0.00023068101010495197, "earlier": 0.00023058456234948988, "first": 0.0002290032372664725, "law": 0.00022896923361557576, "korean": 0.0002286068895001716, "conference": 0.00022847778774013762, "worker": 0.00022828416972148186, "raise": 0.00022777643801001938, "force": 0.000227179989263518, "four": 0.00022674721947831173, "debate": 0.00022556479377929163, "test": 0.00022523499654512, "call": 0.00022504142949674637, "travel": 0.00022467525464252573, "\u2022": 0.00022232109323862595, "general": 0.0002211619811963621, "federal": 0.00021881182037910099, "obamas": 0.00021719836910748713, "top": 0.00021569556361874132, "c": 0.0002155558528615582, "team": 0.0002154064189444724, "international": 0.0002150474417588063, "beijing": 0.00021187012842075498, "conservative": 0.00021178006435733788, "missile": 0.00021114009391046205, "talk": 0.00021057252853625254, "concern": 0.00021001495415685347, "appeal": 0.00020774022431081999, "tillerson": 0.0002072943840440136, "part": 0.00020707607007762105, "island": 0.0002068786474464971, "fund": 0.00020538243599611426, "bill": 0.00020491389745694376, "return": 0.00020481858114764867, "kill": 0.0002025949430108447, "spend": 0.00020130078319923177, "chairman": 0.00019995730725324345, "nuclear": 0.00019989772705325667, "close": 0.00019862754591754212, "rival": 0.00019696219943197583, "request": 0.0001967228111927591, "judge": 0.00019572868725961938, "poll": 0.0001922615237296614, "rohingya": 0.000191846889834017, "offer": 0.00019178621868635895, "critic": 0.00019100191761090807, "ahead": 0.00018941355295295777, "several": 0.00018669426324365208, "step": 0.000186546755141202, "prosecutor": 0.0001861033815947398, "japan": 0.0001860515843678726, "economic": 0.00018597876122051674, "university": 0.0001857185286628345, "deny": 0.00018514082071850084, "work": 0.0001847895363727087, "london": 0.00018402646894611797, "france": 0.00018374929205157705}
